search: convert search query to lowercase for name
Currently if someone searches a name with an uppercase character on
HNScan it will bring them to an error page. Although this is technically
correct behavior, I have run into this issue multiple times where my
phone will autocorrect to a capital. I think by converting to lowercase
which is a valid HNS name, and then serving the user those results makes
for an easier UI/UX for everyone involved.