util: account for airdrops in getBlockTotalFees()
Not all coinbase outputs are miner subsidy payouts. An easy way to
tell is if the output has a corresponding input at the same index.
Besides index 0 (the reuired coinbase input and output) the presence
of an input and witness indicates an airdrop or claim proof.

Additional (optional) lines are commented out to expand this function
in the future if sepcific proof data is desired for the view.