hooks: move to ucl
This commit also adds a new hook for using a "routed modal". Essentially
what this is, is a modal that also triggers some change in URL state. So
the hook accepts a string and then uses that string to indicate that the
modal is open behind a hash. For example, passing in the string of
"connect" would mean that the modal is open when the URL looks like
this: "MYURL/path#connect". This hook was added in this PR rather than
the connect modal PR to allow for the hook folder to remain in the repo.